hackathon/pyproject.toml
[tool.poetry]
name = "hackathon"
version = "0.1.0"
description = ""
authors = ["lukas goldschmidt <lukasgoldschmidt@posteo.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nicegui = "^1.4.15"
pandas = "^2.2.0"
langchain = "^0.1.6"
openai = "^1.12.0"
python-dotenv = "^1.0.1"
langchain-community = "^0.0.19"


[tool.poetry.group.dev.dependencies]
black = "^24.1.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest = "^8.0.0"
pre-commit = "^3.6.0"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
